Cleanup test_test.py, move unrelated tests in test_misc.py
test plan:
make e2e
