+++ cli/tests/e2e/snapshots/test_misc/test_cli_test_show_supported_languages/results.txt
