+++ cli/tests/e2e/snapshots/test_misc/test_cli_test_time/results.txt
