+++ cli/tests/e2e/test_test.py
@@ -2,7 +2,6 @@
 # See https://semgrep.dev/docs/writing-rules/testing-rules/
 # See also test_fixtest.py for the autofix + 'semgrep test' tests.
 import os
-import re
 import subprocess
 from pathlib import Path
 
@@ -11,13 +10,6 @@ from tests.fixtures import RunSemgrep
 from tests.semgrep_runner import SEMGREP_BASE_SCAN_COMMAND
 
 from semgrep.constants import OutputFormat
-
-
-FLOATS = re.compile("([0-9]+).([0-9]+)")
-
-
-def _mask_floats(text_output: str) -> str:
-    return re.sub(FLOATS, "x.xxx", text_output)
 
 
 @pytest.mark.kinda_slow
@@ -32,40 +24,6 @@ def test_cli_test_basic(run_semgrep_in_tmp: RunSemgrep, snapshot):
     snapshot.assert_match(
         results,
         "results.json",
-    )
-
-
-# TODO: I don't understand why this pass
-@pytest.mark.kinda_slow
-def test_cli_test_verbose(run_semgrep_in_tmp: RunSemgrep, snapshot):
-    results, _ = run_semgrep_in_tmp(
-        "rules/cli_test/basic/",
-        options=["--verbose"],
-        target_name="cli_test/basic/",
-        output_format=OutputFormat.TEXT,
-        force_color=True,
-    )
-
-    snapshot.assert_match(
-        _mask_floats(results),
-        "results.txt",
-    )
-
-
-@pytest.mark.osemfail
-@pytest.mark.kinda_slow
-def test_cli_test_time(run_semgrep_in_tmp: RunSemgrep, snapshot):
-    results, _ = run_semgrep_in_tmp(
-        "rules/cli_test/basic/",
-        options=["--time"],
-        target_name="cli_test/basic/",
-        output_format=OutputFormat.TEXT,
-        force_color=True,
-    )
-
-    snapshot.assert_match(
-        _mask_floats(results),
-        "results.txt",
     )
 
 
@@ -95,21 +53,6 @@ def test_cli_test_yaml_language(run_semgrep_in_tmp: RunSemgrep, snapshot):
     snapshot.assert_match(
         results,
         "results.json",
-    )
-
-
-@pytest.mark.kinda_slow
-def test_cli_test_show_supported_languages(run_semgrep_in_tmp: RunSemgrep, snapshot):
-    results, _ = run_semgrep_in_tmp(
-        "rules/cli_test/basic/",
-        options=["--show-supported-languages"],
-        target_name="cli_test/basic/",
-        output_format=OutputFormat.TEXT,
-    )
-
-    snapshot.assert_match(
-        results,
-        "results.txt",
     )
